Divided into separate Debts applications. A new function for displaying lists of contractors  and debts has been created
