+++ apps/debts/choices.py
@@ -1,14 +1,5 @@
 CURRENCIES = [
     ('USD', 'US Dollar'),
     ('EUR', 'Euro'),
-    ('UAH', 'Hryvnia'),
-]
-
-TYPE_CONTRACTOR = [
-    ('МФО', 'Мікрофінансова організація'),
-    ('ФК', 'Факторингова компанія'),
-    ('БУ', 'Банкова установа'),
-    ('СК', 'Страхова компанія'),
-    ('ТК', 'Телекомунікаційна компанія'),
-    ('ЖКГ', 'Житлово-комунальне гопподарство')
+    ('UAH', 'Гривня'),
 ]
