+++ apps/debts/urls.py
@@ -1,10 +1,7 @@
 from django.urls import path
-from .views import list_clients, list_contractors, list_contractor_managers, list_debts
+from .views import ListDebtsView
 
 
 urlpatterns = [
-    path('list_clients/', list_clients),
-    path('list_contractors/', list_contractors),
-    path('list_contractor_managers/', list_contractor_managers),
-    path('list_debts/', list_debts),
+    path('list_debts/', ListDebtsView.as_view(), name='list-debts'),
 ]
