+++ templates/index.html
@@ -3,5 +3,5 @@
     ADA Collection
 {% endblock %}
 {% block content %}
-
+    <h1 align="center">Ця сторінка є тестовою.</h1>
 {% endblock %}
